Test git in idea
