+++ practic/src/main/java/inreco/vlgu/practic/controllers/TestController.java
@@ -47,20 +47,4 @@ public class TestController {
     public String managerAccess() {
         return "Manager Board.";
     }
-
-    @GetMapping("/products")
-    // @PreAuthorize("hasRole('manager')")
-    public ResponseEntity<ProductResponse> products() {
-        return ResponseEntity.ok(new ProductResponse(productService.getProducts()));
-    }
-
-    @PostMapping("/product/create")
-    public ResponseEntity<MessageResponse> createProduct(@Valid @RequestBody ProductCreateRequest productCreateRequest) {
-        if(productService.createProduct(productCreateRequest))
-            return ResponseEntity.ok(new MessageResponse("Service created successfully!"));
-        else
-            return ResponseEntity
-                    .badRequest()
-                    .body(new MessageResponse("Error: Something went wrong("));
-    }
 }
